+++ AppCenterAnalytics/AppCenterAnalytics/include/MSConstants+Flags.h
@@ -1,0 +1,1 @@
+../../../AppCenter/AppCenter/MSConstants+Flags.h