+++ AppCenterCrashes/AppCenterCrashes/Internals/Util/MSApplicationForwarder.m
@@ -24,7 +24,11 @@ static BOOL isApplicationForwarderEnabled() {
 static NSString *const kMSCrashOnExceptionsKey = @"NSApplicationCrashOnExceptions";
 
 static BOOL isCrashOnExceptionsEnabled() {
-  NSNumber *crashOnExceptions = [MS_USER_DEFAULTS objectForKey:kMSCrashOnExceptionsKey];
+
+  // We use NSUserDefaults here instead of MSAppCenterUserDefaults, because
+  // we should use system user defaults for system keys.
+  // MSAppCenterUserDefaults prepends all the keys with "MSAppCenter" prefix.
+  NSNumber *crashOnExceptions = [[NSUserDefaults standardUserDefaults] objectForKey:kMSCrashOnExceptionsKey];
   return [crashOnExceptions boolValue];
 }