+++ Sasquatch/Sasquatch/ViewControllers/MSCrashesViewController.swift
@@ -229,9 +229,10 @@ class MSCrashesViewController: UITableViewController, UIImagePickerControllerDel
   private func pokeAllCrashes() {
     var count = UInt32(0)
     let classList = objc_copyClassList(&count)
+    let classes = UnsafeBufferPointer(start: classList, count: Int(count))
     MSCrash.removeAllCrashes()
     for i in 0..<Int(count){
-      let className: AnyClass = classList![i]
+      let className: AnyClass = classes[i]
       if class_getSuperclass(className) == MSCrash.self && className != MSCrash.self {
         MSCrash.register((className as! MSCrash.Type).init())
       }