+++ SasquatchMac/SasquatchMacSwift/CrashLoader.swift
@@ -34,14 +34,15 @@ class CrashLoader {
     }
     
     private static func pokeAllCrashes() {
-        var count = UInt32(0)
-        let classList = objc_copyClassList(&count)
-        MSCrash.removeAllCrashes()
-        for i in 0..<Int(count) {
-            let className: AnyClass = classList![i]
-            if class_getSuperclass(className) == MSCrash.self && className != MSCrash.self {
-                MSCrash.register((className as! MSCrash.Type).init())
-            }
+      var count = UInt32(0)
+      let classList = objc_copyClassList(&count)
+      let classes = UnsafeBufferPointer(start: classList, count: Int(count))
+      MSCrash.removeAllCrashes()
+      for i in 0..<Int(count){
+        let className: AnyClass = classes[i]
+        if class_getSuperclass(className) == MSCrash.self && className != MSCrash.self {
+          MSCrash.register((className as! MSCrash.Type).init())
         }
+      }
     }
 }